:key: Added Key
Trying to reuse the overlap function as a collision detector. It doesn't work too well. Currently as long as you straddle the width of the key the ghost can got up from the bottom (or down from the top) and be all over the key. Don't know how to fix, but there is some collision